refactor: DRY up repetitive scope-tracking code
Seeing as we're now doing it in 4 places, let's extract it into a
`withScope()` helper.
